cutout base class
