+++ lib/ruboty/yasumimasu.rb
@@ -1,14 +1,16 @@
 # frozen_string_literal: true
 
-require 'google_drive'
 require 'google/api_client'
-require 'google/api_client/client_secrets'
 require 'google/api_client/auth/installed_app'
+require 'google/api_client/client_secrets'
+require 'google_drive'
 
-require 'ruboty/handlers/yasumimasu'
-require 'ruboty/actions/google_spreadsheet/day_off'
+require 'ruboty/actions/google_calendar/base'
 require 'ruboty/actions/google_calendar/day_off'
+require 'ruboty/actions/google_spreadsheet/base'
+require 'ruboty/actions/google_spreadsheet/day_off'
 require 'ruboty/google_calendar/client'
 require 'ruboty/google_spreadsheet/client'
 require 'ruboty/google_spreadsheet/spread_sheet'
+require 'ruboty/handlers/yasumimasu'
 require 'ruboty/yasumimasu/version'
